refactor(Store): 패키지 구조 변경
돈을 받는 Cacher와 로또를 자동 발권하는 TicketingMachine을 Store 패키지에 옮김
